Relocate vdb binaries to openvdb_cmd root directory
