+++ openvdb_cmd/vdb_view/Camera.cc
