+++ openvdb_cmd/vdb_view/Camera.h
