+++ openvdb_cmd/vdb_view/ClipBox.cc
