+++ openvdb_cmd/vdb_view/main.cc
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MPL-2.0
 
 #include <openvdb/openvdb.h>
-#include "../viewer/Viewer.h"
+#include "Viewer.h"
 #include <boost/algorithm/string/classification.hpp> // for boost::is_any_of()
 #include <boost/algorithm/string/predicate.hpp> // for boost::starts_with()
 #include <boost/algorithm/string/split.hpp>
