+++ pendingchanges/openvdb_cmd.txt
@@ -1,0 +1,2 @@
+Build:
+    - Relocated OpenVDB binaries to a new openvdb_cmd root directory.
